--- hybin/MegaBox/MegaBox/Sources/ContentView.swift
@@ -1,17 +1,0 @@
-import SwiftUI
-
-public struct ContentView: View {
-    public init() {}
-
-    public var body: some View {
-        Text("Hello, World!")
-            .padding()
-    }
-}
-
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
