--- hybin/MegaBox/MegaBox/Tests/MegaBoxTests.swift
@@ -1,10 +1,0 @@
-import Testing
-@testable import MegaBox
-
-struct MegaBoxTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
-}
