+++ hybin/MegaBox/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
