+++ hybin/MegaBox/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
