+++ hybin/MegaBox/Resources/Assets.xcassets/Contents.json
