+++ hybin/MegaBox/Resources/Assets.xcassets/appleLogin.imageset/Contents.json
