+++ hybin/MegaBox/Resources/Assets.xcassets/background.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "티켓 배경 1.pdf",
+      "filename" : "티켓 배경 4.pdf",
       "idiom" : "universal"
     }
   ],
