+++ hybin/MegaBox/Resources/Assets.xcassets/background.imageset/티켓 배경 4.pdf
