+++ hybin/MegaBox/Resources/Assets.xcassets/kakaoLogin.imageset/Contents.json
