+++ hybin/MegaBox/Resources/Assets.xcassets/meboxLogo.imageset/Contents.json
