+++ hybin/MegaBox/Resources/Assets.xcassets/naverLogin.imageset/Contents.json
