+++ hybin/MegaBox/Resources/Assets.xcassets/umcLogo.imageset/Contents.json
