+++ hybin/MegaBox/Resources/Color.xcassets/Contents.json
