+++ hybin/MegaBox/Resources/Color.xcassets/LoginBackgroundColor.colorset/Contents.json
