+++ hybin/MegaBox/Resources/Color.xcassets/LoginTextBackgroundColor.colorset/Contents.json
