+++ hybin/MegaBox/Resources/Fonts.swift
@@ -9,7 +9,7 @@
 import Foundation
 import SwiftUI
 
-extension Font {
+public extension Font {
     enum Pretend{
         case extraBold
         case bold
