--- hybin/MegaBox/Tuist.swift
@@ -1,3 +1,0 @@
-import ProjectDescription
-
-let tuist = Tuist(project: .tuist())
