--- hybin/MegaBox/mise.toml
@@ -1,2 +1,0 @@
-[tools]
-tuist = "4.76.1"
